requirements.txt
flask
torchvision
vit-pytorch
pillow
packaging 
